refactor: extract channels CLI command into its own module
Move registerChannelsCommand from integrations.ts into a dedicated
channels.ts file, keeping shared gateway helpers in integrations.ts.
